sidecar/requirements.txt
# These are the only NEW dependencies — fastapi, uvicorn, boto3, pydantic
# are already installed as LiteLLM dependencies on the EC2 instance.
psycopg2-binary>=2.9
Pillow>=10.0
httpx>=0.27
